+++ substra/sdk/backends/base.py
@@ -31,6 +31,10 @@ class BaseBackend(abc.ABC):
         raise NotImplementedError
 
     @abc.abstractmethod
+    def archive(self, key, spec, spec_options=None):
+        raise NotImplementedError
+
+    @abc.abstractmethod
     def add_compute_plan_tuples(self, spec, spec_options):
         raise NotImplementedError